+++ src/js/Messages.Embed.js
@@ -9,7 +9,7 @@ var Messages = {
 	cir_down: "Structure unavailable<br/><small>(the Chemical Identifier Resolver is temporarily offline)</small>",
 	
 	compound: "Loading compound...",
-	protein: "Loading protein...",
+	biomolecule: "Loading biomolecule...",
 	crystal: "Loading crystal...",
 	
 	search: "Searching...",
@@ -17,7 +17,7 @@ var Messages = {
 	no_webgl_support: "You cannot use this render engine because your browser doesn't support WebGL, find out how to get it <a class='link' href='http://get.webgl.org/'>here</a>",
 	no_canvas_support: "Your browser doesn't support this web application, try <a class='link' href='//google.com/chrome' title='A modern browser'>Google Chrome</a> instead.",
 	
-	mobile_old_no_proteins: "Proteins cannot be viewed on your device",
+	mobile_old_no_biomolecules: "Biomolecules cannot be viewed on your device",
 	
 	load_fail: "Failed to load structure <small>This could mean the source database is offline</small>",
 	search_fail: "Structure cannot be found",
@@ -31,7 +31,7 @@ var Messages = {
 		/*
 		Valid strings for {what}
 		- compound
-		- protein
+		- biomolecule
 		- crystal
 		*/
 		
@@ -57,7 +57,7 @@ var Messages = {
 		- load_fail
 		- search_fail
 		
-		- mobile_old_no_proteins
+		- mobile_old_no_biomolecules
 		*/
 		
 		$("body").removeClass("progress-cursor");